requirements.txt
streamlit
opencv-python-headless
numpy
pillow
easyocr
torch
torchvision